Changes made by Seb
