+++ web/src/components/hamburger_initial.module.css
@@ -1,0 +1,45 @@
+/* Parent container styles */
+.parent_container {
+    position: relative;
+    /* Ensure the parent container has a non-static position */
+  }
+
+/*Hamburger menu */
+.hamburger_icon {
+    width: 30px;
+    height: 15px;
+    display: flex;
+    flex-direction: column;
+    justify-content: space-between;
+    cursor: pointer;
+    position: relative;
+    top: 20px;
+    /* Adjust this value to move the menu downwards */
+    left: 20px;
+    /* Adjust this value to move the menu rightwards */
+  }
+  
+  .bar {
+    width: 130%;
+    height: 3px;
+    background-color: #000;
+  }
+  
+  .half_bar {
+    width: 90%;
+    height: 3px;
+    background-color: #000;
+  }
+  
+  /* Open state for hamburger icon */
+  .hamburger-icon.open .bar:first-child {
+    transform: translateY(6px) rotate(45deg);
+  }
+  
+  .hamburger-icon.open .bar:nth-child(2) {
+    opacity: 0;
+  }
+  
+  .hamburger-icon.open .bar:last-child {
+    transform: translateY(-6px) rotate(-45deg);
+  }
