+++ web/src/index.css
@@ -13,7 +13,6 @@
 body {
   background-color: #e1ebff;
   /* Set the background color to the desired color */
-
 }
 
 h1 {
@@ -35,52 +34,6 @@ p {
 .paragraph-type-2{
   font-size: 23px;
   color: white;
-}
-
-/* Parent container styles */
-.parent-container {
-  position: relative;
-  /* Ensure the parent container has a non-static position */
-}
-
-/*Hamburger menu */
-.hamburger-icon {
-  width: 30px;
-  height: 15px;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  cursor: pointer;
-  position: relative;
-  top: 20px;
-  /* Adjust this value to move the menu downwards */
-  left: 20px;
-  /* Adjust this value to move the menu rightwards */
-}
-
-.bar {
-  width: 130%;
-  height: 3px;
-  background-color: #000;
-}
-
-.half-bar {
-  width: 90%;
-  height: 3px;
-  background-color: #000;
-}
-
-/* Open state for hamburger icon */
-.hamburger-icon.open .bar:first-child {
-  transform: translateY(6px) rotate(45deg);
-}
-
-.hamburger-icon.open .bar:nth-child(2) {
-  opacity: 0;
-}
-
-.hamburger-icon.open .bar:last-child {
-  transform: translateY(-6px) rotate(-45deg);
 }
 
 /* Styles for the rectangle container */
