--- web/src/pages/pageLayout.tsx
@@ -1,34 +1,0 @@
-//make the placeholders for objects that are on every page
-//eg the stamp counter
-//eg the hamburger menu
-import React, { useState } from 'react';
-//import React from 'react';
-import './utils/index'; // Import the single CSS file for styling
-
-const HamburgerMenu = () => {
-  const [isOpen, setIsOpen] = useState(false);
-
-  const toggleMenu = () => {
-    setIsOpen(!isOpen);
-  };
-
-  return (
-    <div className="hamburger-menu">
-      <div className={`hamburger-icon ${isOpen ? 'open' : ''}`} onClick={toggleMenu}>
-        <div className="bar"></div>
-        <div className="bar"></div>
-        <div className="bar"></div>
-      </div>
-      <div className={`menu-items ${isOpen ? 'open' : ''}`}>
-        <a href="/">Home</a>
-        <a href="/about">About</a>
-        <a href="/contact">Contact</a>
-      </div>
-    </div>
-  );
-};
-
-export default HamburgerMenu;
-
-
-
